requirements.txt
click
requests
py7zr
rarfile
python-dotenv
pyftpdlib
mega.py
sqlite-utils
yara-python
pandas
jinja2
# Optional: For ClamAV AV scanning
# pip install clamd 